+++ src/app/inobeta-ui/ui/material-forms/material-form.module.ts
@@ -3,7 +3,10 @@ import { IbMaterialFormComponent } from './material-form/material-form.component
 import { IbMaterialFormControlComponent, IbFormControlDirective } from './material-form-control/material-form-control.component';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
-import { MatButtonModule, MatFormFieldModule, MatOptionModule, MatSelectModule, MatRadioModule, MatCheckboxModule, MatInputModule, MatDatepickerModule, MatAutocompleteModule, MatIconModule, MatGridListModule } from '@angular/material';
+import { MatButtonModule, MatFormFieldModule, MatOptionModule,
+  MatSelectModule, MatRadioModule, MatCheckboxModule, MatInputModule,
+  MatDatepickerModule, MatAutocompleteModule, MatIconModule,
+  MatGridListModule, MatTooltipModule } from '@angular/material';
 import { TranslateModule } from '@ngx-translate/core';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { IbMatTextboxComponent } from './controls/textbox';
@@ -16,6 +19,7 @@ import { IbMatLabelComponent } from './controls/label';
 import { IbMatTextareaComponent } from './controls/textarea';
 import { IbMatButtonComponent } from './controls/button';
 import { IbMatPaddingComponent } from './controls/padding';
+import { MatMomentDateModule } from '@angular/material-moment-adapter';
 
 const entryComponents = [
   IbMatTextboxComponent,
@@ -55,7 +59,9 @@ const components = [
     MatAutocompleteModule,
     MatIconModule,
     MatGridListModule,
-    FlexLayoutModule
+    FlexLayoutModule,
+    MatMomentDateModule,
+    MatTooltipModule
   ],
   exports: [
     ...components